+++ compose-mds/src/main/java/ch/sbb/compose_mds/beta/switch/SBBSwitch.kt
@@ -2,113 +2,127 @@ package ch.sbb.compose_mds.beta.switch
 
 import SBBTheme
 import android.content.res.Configuration
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.TweenSpec
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.core.animateIntOffsetAsState
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.selection.toggleable
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import ch.sbb.compose_mds.beta.ExperimentalSBBComponent
+import ch.sbb.compose_mds.sbbicons.SBBIcons
+import ch.sbb.compose_mds.sbbicons.Small
+import ch.sbb.compose_mds.sbbicons.small.TickSmall
 import ch.sbb.compose_mds.theme.SBBColors
-import kotlinx.coroutines.launch
-import kotlin.math.roundToInt
+import ch.sbb.compose_mds.theme.SBBConst
 
 @ExperimentalSBBComponent
 @Composable
 fun SBBSwitch(
+    modifier: Modifier = Modifier,
     checked: Boolean,
     onCheckedChange: ((Boolean) -> Unit)?,
-    modifier: Modifier = Modifier,
     enabled: Boolean = true,
     interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
 ) {
-    val darkTheme = isSystemInDarkTheme()
-    Box(
-        modifier = modifier.alpha(
-            if (enabled) 1.0f else .5f,
+    val color by animateColorAsState(
+        targetValue = when (checked) {
+            true -> MaterialTheme.colorScheme.primary
+            false -> SBBColors.graphite
+        }
+    )
+    val knobSize = 26.dp
+    val knobOffset by animateIntOffsetAsState(
+        targetValue = IntOffset(
+            x = with(LocalDensity.current) {
+                when (checked) {
+                    true -> 52.dp - knobSize
+                    false -> 0.dp
+                }.roundToPx()
+            },
+            y = 0,
         ),
+    )
+    val alpha = when (enabled) {
+        true -> 1f
+        false -> 0.5f
+    }
+    val shadow = when (enabled) {
+        true -> 4.dp
+        false -> 0.dp
+    }
+    Box(
+        modifier = modifier
+            .alpha(alpha)
+            .size(DpSize(53.dp, knobSize + 2.dp))
+            .padding(end = 1.dp)
+            .toggleable(
+                enabled = enabled,
+                value = checked,
+                interactionSource = interactionSource,
+                role = Role.Switch,
+                onValueChange = { onCheckedChange?.invoke(it) },
+                indication = null
+            ),
+        contentAlignment = Alignment.CenterStart,
     ) {
-        val minBound = 0f
-        val maxBound = 52f
-        val valueToOffset = remember<(Boolean) -> Float>(minBound, maxBound) {
-            { value -> if (value) maxBound else minBound }
-        }
-        val targetValue = valueToOffset(checked)
-        val offset = remember { Animatable(targetValue) }
-        val scope = rememberCoroutineScope()
-
-        SideEffect {
-            // min bound might have changed if the icon is only rendered in checked state.
-            offset.updateBounds(lowerBound = minBound)
-        }
-
-        DisposableEffect(checked) {
-            if (offset.targetValue != targetValue) {
-                scope.launch {
-                    offset.animateTo(targetValue, TweenSpec(durationMillis = 100))
-                }
-            }
-            onDispose { }
-        }
-
         Box(
             modifier = Modifier
-                .toggleable(
-                    value = checked,
-                    onValueChange = {
-                        onCheckedChange?.invoke(it)
-                    },
-                    enabled = enabled,
-                    role = Role.Switch,
-                    interactionSource = interactionSource,
-                    indication = null,
-                )
-                .width(50.dp)
-                .clip(RoundedCornerShape(50))
-                .background(
-                    if (checked) SBBColors.red
-                    else if (darkTheme) SBBColors.metal
-                    else SBBColors.smoke,
-                ),
+                .fillMaxWidth()
+                .height(20.dp)
+                .background(color = color, shape = CircleShape),
+        )
+        Box(
+            modifier = Modifier
+                .offset { knobOffset }
+                .size(knobSize)
+                .shadow(elevation = shadow, shape = CircleShape)
+                .background(SBBColors.white, shape = CircleShape)
+                .border(border = BorderStroke(width = 1.dp, color = color), shape = CircleShape),
+            contentAlignment = Alignment.Center,
         ) {
-            Box(
-                modifier = Modifier
-                    .offset { IntOffset(offset.value.roundToInt(), 0) }
-                    .shadow(
-                        elevation = 8.dp,
-                        shape = CircleShape,
-                    )
-                    .padding(2.dp)
-                    .clip(CircleShape)
-                    .size(27.dp)
-                    .background(SBBColors.white),
-            )
+            AnimatedVisibility(
+                visible = checked,
+                enter = fadeIn(),
+                exit = fadeOut(),
+            ) {
+                Icon(
+                    imageVector = SBBIcons.Small.TickSmall,
+                    tint = color,
+                    contentDescription = null
+                )
+            }
         }
-
     }
 }
 
@@ -122,28 +136,31 @@ fun SBBSwitchPreview() {
     val secondChecked = remember { mutableStateOf(true) }
     SBBTheme {
         Column(
-            modifier = Modifier.background(
-                if (darkTheme) SBBColors.black else SBBColors.white
-            ),
+            modifier = Modifier
+                .background(if (darkTheme) SBBColors.black else SBBColors.white)
+                .padding(SBBConst.DEFAULT_PADDING),
+            verticalArrangement = Arrangement.spacedBy(SBBConst.HALF_PADDING),
         ) {
-            Row(modifier = Modifier.padding(8.dp)) {
+            Row(
+                horizontalArrangement = Arrangement.spacedBy(SBBConst.HALF_PADDING),
+            ) {
                 SBBSwitch(
                     checked = firstChecked.value,
                     onCheckedChange = { firstChecked.value = it },
                 )
-                Spacer(modifier = Modifier.width(8.dp))
                 SBBSwitch(
                     checked = secondChecked.value,
                     onCheckedChange = { secondChecked.value = it },
                 )
             }
-            Row(modifier = Modifier.padding(8.dp)) {
+            Row(
+                horizontalArrangement = Arrangement.spacedBy(SBBConst.HALF_PADDING),
+            ) {
                 SBBSwitch(
                     checked = firstChecked.value,
                     onCheckedChange = { firstChecked.value = it },
                     enabled = false,
                 )
-                Spacer(modifier = Modifier.width(8.dp))
                 SBBSwitch(
                     checked = secondChecked.value,
                     onCheckedChange = { secondChecked.value = it },