+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+# Unreleased
+
+### Features
+* Added SBBLoadingIndicator
+
 # 0.1.3: Update Icons
 
 ### Features